Split the utils tests into two separate files to reflect better the new file hide_utils
